[Minor] move triton attention kernels into a separate folder (#1379)
