+++ python/sglang/srt/constrained/__init__.py
@@ -13,6 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+"""For constrained decoding."""
+
 import json
 from typing import Dict, Optional, Union
